// 10-oop-sendiri/src/model/character-enum.ts
export enum CharRarity{
    fiveStar = '5-Star Character',
    fourStar = '4-Star Character'
}

export enum CharType{
    playable = 'Playable Character',
    nonPlayable = 'Non Playable Character'
}

export enum CharVision{
    anemo = 'Anemo',
    geo = 'Geo',
    electro = 'Electro',
    pyro = 'Pyro',
    hydro = 'Hydro',
    cryo = 'Cryo',
    dendro = 'Dendro',
    noVision = 'No Vision'
}

export enum CharWeapon{
    sword = 'Sword',
    claymore = 'Claymore',
    bow = 'Bow',
    catalyst = 'Catalyst',
    polearm = 'Polearm'
}